pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hivstats"
version = "1.0.0"
description = "Various HIV statistics"
requires-python = ">=3.11,<3.13"
authors = [
    { name = "British Columbia Centre for Excellence in HIV/AIDS", email = "vmysak@bccfe.ca" },
]
dependencies = [
    # General libs.
    "Levenshtein==0.25.0",
    "matplotlib==3.8.4",
    "pandas==2.3.3",
    "jarowinkler==2.0.1",
    "biopython==1.83.0",
    "cfeintact @ git+https://github.com/cfe-lab/CFEIntact.git@v1.25.0",

    # Jupyter
    "jupyterlab==4.1.5",
    "jupyterlab_server==2.25.4",
    "notebook==7.1.2",
    "voila==0.5.6",
    "ipywidgets==8.1.2",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.sdist]
include = [ "src" ]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[[tool.mypy.overrides]]
module = "ipywidgets.*"
ignore_missing_imports = true
